Speed up the candidate list build
Four warps instead of two once the rows fill the machine, and int32
positions, which halves the largest store.

+++ aiter/ops/triton/_gluon_kernels/gfx950/attention/pa_mqa_logits_mxfp4.py
@@ -217,8 +217,8 @@ def _prepare_candidates_kernel(
 
     # Padding repeats the last legal block, so a short row still walks blocks
     # it owns and row_ends drops the slots.
-    last = ((nblocks - 1) * C).to(tl.int64)
-    pos = tl.where(key == 0x7FFFFFFF, last, key.to(tl.int64) * C)
+    last = (nblocks - 1) * C
+    pos = tl.where(key == 0x7FFFFFFF, last, key * C)
     tl.store(pos_ptr + row * K + cols, pos)
     tail = tl.minimum(C, end - max_id * C)
     tl.store(cu_ptr + row, tl.where(n_valid > 0, (n_valid - 1) * C + tail, 0))
